+++ spec/image_optim.rb
@@ -1,9 +1,0 @@
-# This class stub is required by ImageOptimizaion#run_check in the
-# image_optimization_spec.
-class ImageOptim
-  def initialize(*args)
-  end
-
-  class BinNotFoundError < Exception
-  end
-end